Render macro, protein with binding site and glycan with reusable components and hooks
