+++ app/routes/scenarios.protein-glycan.tsx
@@ -19,11 +19,7 @@ import { LoaderFunctionArgs } from "@remix-run/node";
 import { WORKFLOW_CONFIG_FILENAME } from "~/bartender-client/constants";
 import { ActionButtons, handleActionButton } from "~/scenarios/actions";
 import { parseFormData } from "~/scenarios/schema";
-import { FormDescription } from "~/scenarios/FormDescription";
-import { FormItem } from "~/scenarios/FormItem";
-import { PDBFileInput } from "~/scenarios/PDBFileInput.client";
 import { action as uploadaction } from "./upload";
-import { MoleculeSubForm } from "~/scenarios/MoleculeSubForm.client";
 import { ActPassSelection, countSelected } from "~/scenarios/ActPassSelection";
 import { ClientOnly } from "~/components/ClientOnly";
 import { mustBeAllowedToSubmit } from "~/auth.server";
@@ -33,6 +29,8 @@ import {
 } from "../scenarios/restraints";
 import { FormErrors } from "../scenarios/FormErrors";
 import { ReferenceStructureInput } from "~/scenarios/ReferenceStructureInput";
+import { MacroMoleculeSubForm } from "~/scenarios/MacroMoleculeSubForm.client";
+import { GlycanMoleculeSubForm } from "~/scenarios/GlycanMoleculeSubForm";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
   await mustBeAllowedToSubmit(request);
@@ -262,7 +260,7 @@ export default function ProteinProteinScenario() {
         {() => (
           <form onSubmit={onSubmit}>
             <div className="grid grid-cols-2 gap-6">
-              <MoleculeSubForm
+              <MacroMoleculeSubForm
                 name="protein"
                 legend="Protein"
                 description="In example named data/1LMQ_r_u.pdb"
@@ -275,13 +273,14 @@ export default function ProteinProteinScenario() {
               only active as restraints flavour kind,
               render as ball+stick  */}
               {/* TODO molstar has carbohydrate representation, check if there is equiv in ngl */}
-              <MoleculeSubForm
+              <GlycanMoleculeSubForm
                 name="glycan"
                 legend="Glycan"
                 description="In example named data/1LMQ_l_u.pdb"
                 actpass={glycanActPass}
                 onActPassChange={setGlycanActPass}
                 targetChain="B"
+                preprocessPipeline="delhetatmkeepcoord"
               />
             </div>
             <ReferenceStructureInput>
