+++ app/scenarios/MoleculeSubFormWrapper.tsx
@@ -1,0 +1,20 @@
+import { FormDescription } from "./FormDescription";
+
+export function MoleculeSubFormWrapper({
+  legend,
+  description,
+  children,
+}: {
+  legend: string;
+  description: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <fieldset className="border border-solid border-primary p-3">
+      <legend>{legend}</legend>
+
+      <FormDescription>{description}</FormDescription>
+      {children}
+    </fieldset>
+  );
+}
