+++ app/scenarios/ShowSurfaceBuriedToggles.tsx
@@ -1,6 +1,43 @@
-import { useId } from "react";
+import { useId, useState } from "react";
 import { Label } from "~/components/ui/label";
 import { ToggleGroup, ToggleGroupItem } from "~/components/ui/toggle-group";
+
+export function useShowSurfaceBuriedToggles({
+  surfaceResidues,
+  residues,
+}: {
+  surfaceResidues: number[];
+  residues: number[];
+}) {
+  const [showSurface, setShowSurface] = useState(false);
+  const [showBuried, setShowBuried] = useState(false);
+  let surfaceOrBuriedResidues: number[] = [];
+  if (showSurface) {
+    surfaceOrBuriedResidues = surfaceResidues;
+  } else if (showBuried) {
+    surfaceOrBuriedResidues = residues.filter(
+      (r) => !surfaceResidues.includes(r),
+    );
+  }
+  return {
+    resetShowBuriedToggle(renderedAsSurface: boolean) {
+      if (renderedAsSurface) {
+        setShowBuried(false);
+        setShowSurface(true);
+      } else {
+        setShowSurface(false);
+        setShowBuried(false);
+      }
+    },
+    surfaceOrBuriedResidues,
+    surfaceBuriedTogglesProps: {
+      surface: showSurface,
+      setSurface: setShowSurface,
+      buried: showBuried,
+      setBuried: setShowBuried,
+    },
+  };
+}
 
 export function ShowSurfaceBuriedToggles({
   surface,
@@ -27,7 +64,7 @@ export function ShowSurfaceBuriedToggles({
   }
 
   return (
-    <div className="flex items-center space-x-2">
+    <div className="flex items-center space-x-2 py-2">
       <Label htmlFor={id}>Show</Label>
       <ToggleGroup
         type="single"
